Move measurement to a separate file
