+++ src/qutip_qip/operations/__init__.py
@@ -3,3 +3,4 @@ Operations on quantum circuits.
 """
 from .gates import *
 from .gateclass import *
+from .measurement import *
