+++ tests/test_qasm.py
@@ -3,9 +3,9 @@ import numpy as np
 from pathlib import Path
 
 from qutip_qip.qasm import read_qasm, circuit_to_qasm_str
-from qutip_qip.circuit import Measurement, QubitCircuit
+from qutip_qip.circuit import QubitCircuit
 from qutip import tensor, rand_ket, basis, rand_dm, identity
-from qutip_qip.operations import cnot, ry
+from qutip_qip.operations import cnot, ry, Measurement
 
 
 @pytest.mark.parametrize(["filename", "error", "error_message"], [
